Attempting to modularize existing code
